uint to size_t for portability.

+++ src/DatFile.cpp
@@ -330,7 +330,7 @@ namespace gw2b {
 		return Array<byte>( );
 	}
 
-	DatFile::IdentificationResult DatFile::identifyFileType( const byte* p_data, uint p_size, ANetFileType& po_fileType ) {
+	DatFile::IdentificationResult DatFile::identifyFileType( const byte* p_data, size_t p_size, ANetFileType& po_fileType ) {
 		if ( p_size < 4 ) {
 			po_fileType = ANFT_Unknown; return IR_Failure;
 		}
